Add caret mixin
This is the little triangle guy which makes popups and dropdowns
"belong" to something.

This is the simplest form of those dongles, an :after pseudo element.

In order to add a border to the caret we could overlay it on a
:before element and use some css magic to offset them slightly.

This would be a great thing to demo on a style guide/